chore: try to realize FSD refactor: change all paths to @ alias
