--- src/Header/Header.css
@@ -1,77 +1,0 @@
-.header {
-  background-color: #3f72af;
-  box-sizing: border-box;
-  height: 88px;
-  padding: 24px;
-  color: #fff;
-  display: flex;
-  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.5);
-}
-
-.header-title {
-  font-weight: 400;
-  font-size: 36px;
-  line-height: 40px;
-  margin-right: 65px;
-}
-
-.header-nav {
-  flex-grow: 1;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.header-list {
-  display: flex;
-  list-style-type: none;
-}
-
-.header-list-item {
-  box-shadow: inset 0 -2px 0 0 #dbe2ef;
-}
-
-.header-list-item-active {
-  box-shadow: inset 0 -2px 0 0 #8fb6ff;
-}
-
-.header-list-item:first-child {
-  margin-right: 19px;
-}
-
-.header-list-item a {
-  font-size: 24px;
-  line-height: 2px;
-  color: #dbe2ef;
-  position: relative;
-}
-
-/* .header-list-item a:first-child {
-    border-right: 2px solid #dbe2ef;
-  } */
-
-[href="/tasks"]::after {
-  content: "";
-  box-sizing: border-box;
-  width: 0;
-  height: 28px;
-  border-right: 2px solid #dbe2ef;
-  position: absolute;
-  top: 2px;
-  left: 90px;
-}
-
-.header-list-item-active a {
-  color: #8fb6ff;
-}
-
-.header-button {
-  cursor: pointer;
-  box-sizing: border-box;
-  height: 46px;
-  border: none;
-  color: #fff;
-  font-size: 20px;
-  line-height: 46px;
-  background-color: #3f72af;
-}
