--- src/Lists/Tasks.tsx
@@ -1,21 +1,0 @@
-/* VENDOR */
-import { useSelector } from "react-redux";
-
-/* APPLICATION */
-import { ListItem } from "./ListItem";
-import { selectAllTasks } from "../features/tasksSlice";
-import React from "react";
-
-export const Tasks: React.FC = () => {
-  const tasks = useSelector(selectAllTasks);
-
-  return (
-    <ul>
-      {tasks.map((task) => (
-        <ListItem key={task.id} item={task} />
-      ))}
-    </ul>
-  );
-};
-
-// можно попробовать вынести svg в отдельный файл в виде кода
