--- src/Modal/Modal.css
@@ -1,213 +1,0 @@
-.modal {
-  height: 100vh;
-  width: 100vw;
-  background-color: rgba(0, 0, 0, 0.4);
-  position: fixed;
-  top: 0;
-  left: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  opacity: 0;
-  pointer-events: none;
-  z-index: 1;
-}
-
-.modal.active {
-  opacity: 1;
-  pointer-events: all;
-}
-
-.modal__content {
-  box-sizing: border-box;
-  width: 800px;
-  padding: 24px;
-  background: #dbe2ef;
-  box-shadow: 0px 8px 24px rgba(0, 0, 0, 0.5);
-  border-radius: 4px;
-}
-
-.modal__content.small {
-  width: 600px;
-}
-
-.modal__content-header {
-  display: flex;
-  justify-content: space-between;
-}
-
-.modal__content-title {
-  font-style: normal;
-  font-weight: 500;
-  font-size: 40px;
-  line-height: 56px;
-  color: #3f72af;
-  margin-bottom: 32px;
-}
-
-.modal__content-header__btn {
-  cursor: pointer;
-  box-sizing: border-box;
-  width: 25px;
-  height: 25px;
-  border: none;
-  background: transparent;
-}
-
-.modal__content-text {
-  margin-bottom: 24px;
-  font-style: normal;
-  font-weight: 400;
-  font-size: 20px;
-  line-height: 24px;
-  color: #000000;
-}
-
-.modal__content-footer {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.modal__content_row {
-  display: flex;
-  justify-content: space-between;
-}
-
-.modalinput {
-  box-sizing: border-box;
-  padding: 8px;
-  height: 46px;
-  width: 100%;
-  background: transparent;
-  border: 2px solid #3f72af;
-  border-radius: 4px;
-  font-style: normal;
-  font-weight: 400;
-  font-size: 20px;
-  line-height: 46px;
-}
-
-.modalinput::placeholder {
-  color: rgba(0, 0, 0, 0.5);
-}
-
-.modalinput-wrapper {
-  box-sizing: border-box;
-  width: 364px;
-  position: relative;
-  margin-bottom: 34px;
-}
-
-.modalinput-wrapper.large {
-  width: 100%;
-}
-
-.modalinput:focus {
-  outline: none;
-}
-
-[for="modalinput"],
-[for="modaltextarea"],
-.dropdown-label {
-  font-style: normal;
-  display: inline-block;
-  background-color: #dbe2ef;
-  font-weight: 400;
-  padding: 4px;
-  font-size: 18px;
-  line-height: 20px;
-  color: #3f72af;
-  position: absolute;
-  left: 12px;
-  top: -15px;
-}
-
-.modalinput-icon {
-  position: absolute;
-  left: 55px;
-  top: -11px;
-  z-index: 1;
-}
-
-.modaltextarea-wrapper {
-  position: relative;
-}
-
-.modaltextarea {
-  resize: none;
-  box-sizing: border-box;
-  width: 100%;
-  height: 72px;
-  margin-bottom: 24px;
-  background: transparent;
-  border: 2px solid #3f72af;
-  border-radius: 4px;
-  font-style: normal;
-  font-weight: 400;
-  font-size: 20px;
-  line-height: 26px;
-  padding: 8px;
-}
-
-.modaltextarea::placeholder {
-  color: rgba(0, 0, 0, 0.5);
-}
-
-.modaltextarea:focus {
-  outline: none;
-}
-
-.dropdown {
-  cursor: pointer;
-  box-sizing: border-box;
-  width: 364px;
-  height: 24px;
-  position: relative;
-  z-index: 1;
-}
-
-.dropdown .dropdown-btn {
-  box-sizing: border-box;
-  padding: 8px;
-  display: flex;
-  align-items: center; /* для треугольника */
-  justify-content: space-between;
-  background: #dbe2ef;
-  border: 2px solid #3f72af;
-  border-radius: 4px;
-  font-style: normal;
-  font-weight: 400;
-  font-size: 20px;
-  height: 46px;
-  color: #000;
-}
-
-.dropdown .dropdown-btn.placeholder {
-  color: rgba(0, 0, 0, 0.5);
-}
-
-.dropdown .dropdown-content {
-  box-sizing: border-box;
-  position: absolute;
-  top: 45px;
-  background: #dbe2ef;
-  border: 2px solid #3f72af;
-  border-radius: 4px;
-  width: 364px;
-}
-
-.dropdown .dropdown-content .dropdown-item {
-  position: relative;
-  left: -2px;
-  box-sizing: border-box;
-  padding: 10px;
-  width: 364px;
-  border: none;
-  border-radius: 4px;
-  color: #3f72af;
-}
-
-.dropdown .dropdown-content .dropdown-item:hover {
-  background-color: #3f72ff;
-  color: #fff;
-}
