--- src/Modal/Modal.tsx
@@ -1,39 +1,0 @@
-/* VENDOR */
-import React from "react";
-
-/* APPLICATION */
-import "./Modal.css";
-
-interface ModalProps {
-  item?: {
-    id: string;
-    name: string;
-    description: string;
-    category?: string;
-  };
-  active: boolean;
-  setActive: React.Dispatch<React.SetStateAction<boolean>>;
-  children: React.ReactNode;
-  clearState?(): void;
-}
-
-export const Modal: React.FC<ModalProps> = ({
-  clearState,
-  active,
-  setActive,
-  children,
-}) => {
-  return (
-    <div
-      className={active ? "modal active" : "modal"}
-      onClick={() => {
-        clearState && clearState();
-        setActive(false);
-      }}
-    >
-      <div className="modal__content" onClick={(e) => e.stopPropagation()}>
-        {children}
-      </div>
-    </div>
-  );
-};
