--- src/Modal/ModalBtn.css
@@ -1,26 +1,0 @@
-.modalbtn {
-  box-sizing: border-box;
-  height: 46px;
-  width: 120px;
-  font-weight: 400;
-  font-size: 20px;
-  line-height: 46px;
-  color: #3f72af;
-  border: 2px solid #3f72af;
-  background: transparent;
-  border-radius: 4px;
-  cursor: pointer;
-}
-
-.modalbtn:first-child {
-  margin-right: 24px;
-}
-
-.modalbtn.primary {
-  color: #fff;
-  background-color: #3f72af;
-}
-
-.modalbtn.primary.large {
-  width: 200px;
-}
