--- src/Modal/ModalRow.tsx
@@ -1,24 +1,0 @@
-import { ModalInput } from "./ModalInput";
-import { ModalDropdown } from "./ModalDropdown";
-import React from "react";
-
-interface ModalRowProps {
-  name: string;
-  setName: React.Dispatch<React.SetStateAction<string>>;
-  selected: string | undefined;
-  setSelected: React.Dispatch<React.SetStateAction<string>>;
-}
-
-export const ModalRow: React.FC<ModalRowProps> = ({
-  name,
-  setName,
-  selected,
-  setSelected,
-}) => {
-  return (
-    <div className="modal__content_row">
-      <ModalInput name={name} setName={setName} />
-      <ModalDropdown selected={selected} setSelected={setSelected} />
-    </div>
-  );
-};
