+++ src/app/custom.d.ts
@@ -1,0 +1,8 @@
+declare module "*.svg" {
+    const content: any;
+    export default content;
+}
+declare module "*.css" {
+    const content: any;
+    export default content;
+}
