+++ src/app/styles/App.css
@@ -1,0 +1,71 @@
+html {
+  background-color: #dbe2ef;
+}
+
+/* Lists */
+.list-item {
+  box-sizing: border-box;
+  padding: 10px;
+  margin: auto;
+  display: flex;
+  justify-content: space-between;
+  width: 1000px;
+  height: 80px;
+  border-bottom: 2px solid #3f72af;
+}
+
+.list-item:hover {
+  background-color: rgba(37, 127, 234, 0.2);
+}
+
+.list-item-col1 {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.list-item-col1-row1 {
+  display: flex;
+  align-items: center;
+}
+
+.list-item-col1-row1__title {
+  font-weight: 500;
+  font-size: 18px;
+  line-height: 20px;
+  color: #000000;
+  margin-right: 16px;
+}
+
+.list-item-col1-row1__category {
+  font-weight: 400;
+  font-size: 18px;
+  color: #3f72af;
+}
+
+.list-item-col1-row1__category::before {
+  content: url("@/shared/icons/folder.svg");
+  margin-right: 4px;
+  position: relative;
+  top: 3px;
+}
+
+.list-item-col1-row2 {
+  color: #808080;
+  font-weight: 400;
+  font-size: 18px;
+  line-height: 20px;
+}
+
+.list-item-col2__btn {
+  cursor: pointer;
+  box-sizing: border-box;
+  width: 25px;
+  height: 25px;
+  border: none;
+  background: transparent;
+}
+
+.list-item-col2__btn:first-child {
+  margin-right: 15px;
+}
