+++ src/index.tsx
@@ -5,10 +5,10 @@ import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
 
 /* APPLICATION */
-import { store } from "./app/store";
-import App from "./App";
+import { store } from "@/app/redux/store";
+import App from "@/app/App";
 //import reportWebVitals from './reportWebVitals';
-import "./index.css";
+import "@/app/styles/index.css";
 
 const container = document.getElementById("root")!;
 const root = createRoot(container);
