+++ src/pages/Categories/ui/Categories.tsx
@@ -1,0 +1,18 @@
+/* VENDOR */
+import React from "react";
+
+/* APPLICATION */
+import {ListItem} from "@/shared/ui/ListItem";
+import {useAppSelector} from "@/shared/hooks/hooks";
+
+export const Categories = () => {
+    const categories = useAppSelector(state => state.categories);
+
+    return (
+        <ul>
+            {categories.map((category) => (
+                <ListItem key={category.id} item={category} />
+            ))}
+        </ul>
+    );
+};
