+++ src/widgets/Header/ui/Header.module.css
@@ -1,0 +1,77 @@
+.header {
+  background-color: #3f72af;
+  box-sizing: border-box;
+  height: 88px;
+  padding: 24px;
+  color: #fff;
+  display: flex;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.5);
+}
+
+.title {
+  font-weight: 400;
+  font-size: 36px;
+  line-height: 40px;
+  margin-right: 65px;
+}
+
+.nav {
+  flex-grow: 1;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.list {
+  display: flex;
+  list-style-type: none;
+}
+
+.listItem {
+  box-shadow: inset 0 -2px 0 0 #dbe2ef;
+}
+
+.listItem:first-child {
+  margin-right: 19px;
+}
+
+.listItem a {
+  font-size: 24px;
+  line-height: 2px;
+  color: #dbe2ef;
+  position: relative;
+}
+/* .listItem a:first-child {
+    border-right: 2px solid #dbe2ef;
+  } */
+
+[href="/tasks"]::after {
+  content: "";
+  box-sizing: border-box;
+  width: 0;
+  height: 28px;
+  border-right: 2px solid #dbe2ef;
+  position: absolute;
+  top: 2px;
+  left: 90px;
+}
+
+.active {
+  box-shadow: inset 0 -2px 0 0 #8fb6ff;
+}
+
+
+.active a {
+  color: #8fb6ff;
+}
+
+.button {
+  cursor: pointer;
+  box-sizing: border-box;
+  height: 46px;
+  border: none;
+  color: #fff;
+  font-size: 20px;
+  line-height: 46px;
+  background-color: #3f72af;
+}
